--- internal/keygen/generator.go
@@ -1,17 +1,0 @@
-package keygen
-
-import (
-	"math/rand"
-	"time"
-)
-
-func RandomStr() string {
-	const letterBytes = "0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ"
-	n := 6
-	bts := make([]byte, n)
-	rand.Seed(time.Now().UnixNano())
-	for i := 0; i < n; i++ {
-		bts[i] = letterBytes[rand.Intn(len(letterBytes))]
-	}
-	return string(bts)
-}
